solved door animation
A few minutes ago I had the problem that my animation works but the door opjects did not move. I noticed that I had marked the doors as static. I made the left and right door non- static. I solved the problem and now if hitting "E" when in frotn of red keypad -> door will open and close.