added collect to inventory
have now scripts to collect collection items. Each gets a script with name and image and can be made available as collection item by attaching the script to the object. object gets collected if player collides with it. object gets deactivated and image for object gets enabled in inventory panel